split r4-auth into r4-sign{up,in,out} components
