+++ src/components/r4-sign-in.js
@@ -1,0 +1,20 @@
+import { html } from 'lit'
+import { signIn } from '@radio4000/sdk'
+
+const R4SignIn = () => {
+	async function handleSignIn(event) {
+		event.preventDefault()
+		const fd = new FormData(event.target)
+		const { data, error } = await signIn({ email: fd.get('email'), password: fd.get('password') })
+		console.log(data, error)
+	}
+	return html`
+		<form @submit=${handleSignIn}>
+			<input name="email" type="email" required value="oskar@rough.dk" />
+			<input name="password" type="password" required />
+			<button type="submit">Sign In</button>
+		</form>
+	`
+}
+
+export default R4SignIn
