+++ src/components/r4-sign-out.js
@@ -1,0 +1,10 @@
+import { html } from 'lit'
+import { signOut } from '@radio4000/sdk'
+
+const R4SignOut = () => {
+	return html`
+		<button type="button" @click=${() => signOut()}>Sign Out</button>
+	`
+}
+
+export default R4SignOut
